requirements.txt
psycopg2-binary
pandas
scipy
pika
requests